Add a temporary fix failing queryset filters in nested routes
If a route such as /users/NOT_AN_INTEGER/settings would be
requested and the /users endpoint expects an integer as the second
part of the url then this request would throw a internal server error.

This is a temp fix and a proper solution should be found. Discussion
about this can be found on Github:
https://github.com/chibisov/drf-extensions/issues/86
https://github.com/chibisov/drf-extensions/pull/63
https://github.com/chibisov/drf-extensions/issues/50